Moving tests folder because I can't exclude a top level package
