+++ pyTEM/lib/micro_ed/add_basf_icon_to_tkinter_window.py
@@ -15,5 +15,7 @@ def add_basf_icon_to_tkinter_window(root: tk.Tk) -> None:
         The tkinter window of which you want to add the BASF icon.
     :return: None.
     """
-    path_to_ico = os.path.join(BASEDIR, "ico/BASF.ico")
+    path_to_ico = os.path.join(BASEDIR, "ico\\BASF.ico")
+    print("Path to ico: " + str(path_to_ico))
+    print(type(path_to_ico))
     root.iconbitmap(path_to_ico)
