+++ pyTEM/test/interface/hyperspy_test.py
