+++ pyTEM/test/interface/investigate_test_images.py
