+++ pyTEM/test/interface/tif_stack_testing.py
