+++ pyTEM_scripts/bulk_carbon_analysis.py
@@ -1,0 +1,12 @@
+
+
+def bulk_carbon_analysis():
+    """
+
+    :return:
+    """
+    raise NotImplementedError
+
+
+if __name__ == "__main__":
+    bulk_carbon_analysis()
